fatafat 218 - MetadataAPI interface
The methods and the methods that implement or interact with them all have the USER ID
on the method signatures. The argument added is userid : Option[String] with initialization
when possible.

ScalaDoc headers were built for all of the MetadataAPIImpl as well.  There are missing
descriptions, but the basic structure is there so as the code is modified it will not be
difficult to amend the documentation for the API docs.

No work changing the interfaces to use the MdMgr yet instead of passing in jpmml peculiar things.
We need to discuss this, Krishna.  Ping me when you get a chance.