Fatafat 218 JpmmlAdapter get MdMgr
Highlights:
1. More documentaiton cleanup in the MetadataAPIImpl
2. Fixed RemoveConcept handling in ConceptService
3. ModelInfo and ModelContext both contain the namespace.name and version (string form) in order
to support Model factories that manage more than one model (e.g., JpmmlAdapter).  We need to know
which model def the engine is interested.  The JPMMLInfo is no longer there, per discussion on thread
with Krishna.
4. IsValidMessage in ModelBaseObj now sends the namespace.name and version of the model of interest
that is looked up in the JpmmlAdapter using the MdMgr passed at ModelBaseObject initialization time
in the KamanjaMetadata model.
5. The JpmmlAdapter now caches instances of the models with a ModelMessageThread key.  This means
that there will be one instance of the model for each thread in use for the model execution management
on the node.

I will add some logging to the JpmmlAdapter object (factory) and then begin testing.  There are three
models currently that all use the Iris dataset.  I will add that to the project so we can have a
test suite for the three models.